Update config.php
SQL Fehler behoben

+++ application/modules/admin/config/config.php
@@ -839,10 +839,12 @@ class Config extends \Ilch\Config\Install
                     ->execute()
                     ->fetchList();
 
-                $orphanedRows = array_diff($idsArticlesContent, $idsArticles);
-                $this->db()->delete()->from('articles_content')
-                    ->where(['article_id' => $orphanedRows])
-                    ->execute();
+                $orphanedRows = array_diff($idsArticlesContent ?? [], $idsArticles ?? []);
+                if (count($orphanedRows) > 0) {
+                    $this->db()->delete()->from('articles_content')
+                        ->where(['article_id' => $orphanedRows])
+                        ->execute();
+                }
 
                 $this->db()->query('ALTER TABLE `[prefix]_articles_content` ADD CONSTRAINT `FK_[prefix]_articles_content_[prefix]_articles` FOREIGN KEY (`article_id`) REFERENCES `[prefix]_articles` (`id`) ON UPDATE NO ACTION ON DELETE CASCADE;');
